+++ src/compas/geometry/primitives/polyline.py
@@ -344,6 +344,13 @@ class Polyline(Primitive):
         list
             list[:class:`~compas.geometry.Point`]
 
+        Examples
+        --------
+        >>> polyline = Polyline([(0, 0, 0), (1, 1, 0), (2, 3, 0), (4, 4, 0), (5, 2, 0)])
+        >>> divided_polylines = polyline.divide(3)
+        >>> divided_polyline
+        [Point(0.000, 0.000, 0.000), Point(1.578, 2.157, 0.000), Point(3.578, 3.789, 0.000), Point(5.000, 2.000, 0.000)]
+
         """
         segment_length = self.length / num_segments
         return self.divide_polyline_by_length(segment_length, False)
@@ -351,7 +358,7 @@ class Polyline(Primitive):
     divide = divide_polyline
 
     def divide_polyline_by_length(self, length, strict=True, tol=1e-06):
-        """Splits a polyline in segments of a given length.
+        """Divide a polyline in segments of a given length.
 
         Parameters
         ----------
@@ -370,6 +377,18 @@ class Polyline(Primitive):
         -----
         The points of the new polyline are constrained to the segments of the old polyline.
         However, since the old points are not part of the new set of points, the geometry of the polyline will change.
+
+        Examples
+        --------
+        >>> polyline = Polyline([(0, 0, 0), (1, 1, 0), (2, 3, 0), (4, 4, 0), (5, 2, 0)])
+        >>> divided_polylines = polyline.divide_by_length(3)
+        >>> divided_polyline
+        [Point(0.000, 0.000, 0.000), Point(1.709, 2.418, 0.000), Point(4.051, 3.898, 0.000)]
+
+        >>> polyline = Polyline([(0, 0, 0), (1, 1, 0), (2, 3, 0), (4, 4, 0), (5, 2, 0)])
+        >>> divided_polylines = polyline.divide_by_length(3, strict=False)
+        >>> divided_polyline
+        [Point(0.000, 0.000, 0.000), Point(1.709, 2.418, 0.000), Point(4.051, 3.898, 0.000), Point(5.000, 2.000, 0.000)]
 
         """
         num_pts = int(self.length / length)
@@ -403,6 +422,105 @@ class Polyline(Primitive):
         return division_pts
 
     divide_by_length = divide_polyline_by_length
+
+    def split_polyline_by_length(self, length, strict=True):
+        """Split a polyline in segments of a given length.
+
+        Parameters
+        ----------
+        length : float
+            Length of the segments.
+        strict : bool, optional
+            If False, the remainder segment will be added even if it is smaller than the desired length
+        tol : float, optional
+            Floating point error tolerance.
+
+        Returns
+        -------
+        list[:class:`~compas.geometry.Polyline`]
+
+        Examples
+        --------
+        >>> from compas.geometry import Polyline
+        >>> polyline = Polyline([(0, 0, 0), (1, 1, 0), (2, 3, 0), (4, 4, 0), (5, 2, 0)])
+        >>> split_polylines = polyline.split_polyline_by_length(3)
+        >>> split_polylines
+        [Polyline([Point(0.000, 0.000, 0.000), Point(1.000, 1.000, 0.000), Point(1.709, 2.418, 0.000)]),\
+        Polyline([Point(1.709, 2.418, 0.000), Point(2.000, 3.000, 0.000), Point(4.000, 4.000, 0.000),\
+        Point(4.051, 3.898, 0.000)])]
+
+        >>> from compas.geometry import Polyline
+        >>> polyline = Polyline([(0, 0, 0), (1, 1, 0), (2, 3, 0), (4, 4, 0), (5, 2, 0)])
+        >>> split_polylines = polyline.split_polyline_by_length(3, strict=False)
+        >>> split_polylines
+        [Polyline([Point(0.000, 0.000, 0.000), Point(1.000, 1.000, 0.000), Point(1.709, 2.418, 0.000)]),\
+        Polyline([Point(1.709, 2.418, 0.000), Point(2.000, 3.000, 0.000), Point(4.000, 4.000, 0.000),\
+        Point(4.051, 3.898, 0.000)]), Polyline([Point(4.051, 3.898, 0.000), Point(5.000, 2.000, 0.000)])]
+
+        """
+        if length <= 0:
+            raise ValueError("Length should be bigger than 0.")
+        elif length > self.length:
+            raise ValueError("Polyline length {0} is smaller than input length {1}.".format(self.length, length))
+        divided_polylines = []
+        polyline_copy = self.copy()
+        segment = Polyline([self[0]])  # Start a new segment
+        i, current_length = 0, 0
+        polyline_points_num = len(polyline_copy)
+        while i < polyline_points_num - 1:
+            pt1, pt2 = polyline_copy.points[i : i + 2]
+            line_length = pt1.distance_to_point(pt2)
+            current_length += line_length
+            if current_length <= length:
+                segment.points.append(pt2)
+                i += 1
+            else:
+                amp = 1 - ((current_length - length) / line_length)
+                new_pt = pt1 + (pt2 - pt1).scaled(amp)
+                polyline_copy.points.insert(i + 1, new_pt)
+                segment.points.append(new_pt)
+                divided_polylines.append(segment)
+                segment = Polyline([new_pt])  # Start a new segment
+                current_length = 0
+                i += 1
+                polyline_points_num = len(polyline_copy)
+        if not strict and len(divided_polylines):
+            divided_polylines.append(segment)  # Add the last segment
+        return divided_polylines
+
+    split_by_length = split_polyline_by_length
+
+    def split_polyline(self, num_segments):
+        """Split a polyline in equal segments.
+
+        Parameters
+        ----------
+        num_segments : int
+
+        Returns
+        -------
+        list
+            list[:class:`~compas.geometry.Polyline`]
+
+        Examples
+        --------
+        >>> from compas.geometry import Polyline
+        >>> polyline = Polyline([(0, 0, 0), (1, 1, 0), (2, 3, 0), (4, 4, 0), (5, 2, 0)])
+        >>> split_polylines = polyline.split_polyline(3)
+        >>> split_polylines
+        [Polyline([Point(0.000, 0.000, 0.000), Point(1.000, 1.000, 0.000), Point(1.578, 2.157, 0.000)]),\
+        Polyline([Point(1.578, 2.157, 0.000), Point(2.000, 3.000, 0.000), Point(3.578, 3.789, 0.000)]),\
+        Polyline([Point(3.578, 3.789, 0.000), Point(4.000, 4.000, 0.000), Point(5.000, 2.000, 0.000)])]
+        """
+        if num_segments < 1:
+            raise ValueError("Number of segments must be greater than or equal to 1.")
+        elif num_segments == 1:
+            return [self]
+        total_length = self.length
+        segment_length = total_length / num_segments
+        return self.split_polyline_by_length(segment_length, False)
+
+    split = split_polyline
 
     def extend(self, length):
         """Extends a polyline by a given length, by modifying the first and/or last point tangentially.