+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 * Added `create_id` to `compas_ghpython.utilities`. (moved from `compas_fab`)
 * Added representation for features in `compas.datastructures.Part`.
+* Added  `split` and `split_by_length` to `compas.geometry.Polyline`.
 * Added `compas.rpc.XFunc`.
 * Added `compas.data.Data.validate_jsonstring`.
 * Added `compas.data.Data.validate_jsondata`.
@@ -62,6 +63,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 * Fixed bug that caused a new-line at the end of the `compas.HERE` constant in IronPython for Mac.
+* Fixed unbound method usage of `.cross()` on `Plane`, `Vector` and `Frame`.
 * Fixed Grasshopper `draw_polylines` method to return `PolylineCurve` instead of `Polyline` because the latter shows as only points.
 * Fixed `area_polygon` that was, in some cases, returning a negative area.
 * Fixed uninstall post-process.